requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.5.0
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
appdirs==1.4.4
APScheduler==3.11.0
argcomplete==3.5.3
attrs==23.2.0
autocommand==2.2.2
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.2.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cssselect==1.3.0
distro==1.9.0
fake-useragent==2.1.0
filelock==3.18.0
frozenlist==1.4.1
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
lxml==5.3.1
lxml_html_clean==0.4.1
magic-filter==1.0.12
Mako==1.3.9
Markdown==3.7
MarkupSafe==2.1.5
meson==1.7.0
more-itertools==10.6.0
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
pipx==1.7.1
platformdirs==4.3.6
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==11.1.1
pygdbmi==0.11.0.0
pyppeteer==2.0.0
pyproject_hooks==1.2.0
pyquery==2.0.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-html==0.10.0
semantic_version==2.10.0
sentry-sdk==2.23.1
setuptools==75.8.0
setuptools-rust==1.10.2
setuptools-scm==8.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
TBB==2022.0.0
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.12.2
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==1.26.20
userpath==1.9.2
w3lib==2.3.1
websocket-client==1.8.0
websockets==10.4
wheel==0.45.0
wsproto==1.2.0
yarl==1.13.1
zipp==3.21.0
